// src/app/main/management-tools/tabs/components/building-repository/model/building-repo.model.ts
export class BuildingRepo {
    "id": number;
    "clientOrganisationId": number;
    "color": string;
    "comments": string;
    "contactEmail": string;
    "contactNumber": number;
    "createdBy": number;
    "dateOfLien": string;
    "emergencyContactEmail": string;
    "emergencyContactName": string;
    "emergencyContactNumber": number;
    "firstName": string;
    "handicapInfo": string;
    "lastName": string;
    "licenseNumber": string;
    "lienType": number;
    "personTenantType": number;
    "petDescription": string;
    "petName": string;
    "tenantStatus": number;
    "unitNumber": string;
    "unitType": number;
    "vehicleMode": string;
    "createdAt": string;

    "unitTypeDisplayText": string;
    "tenantStatusDisplayText": string;
    "lienTypeDisplayText": string;
    "lienDisplayDate": string;
    "personTenantTypeDisplayText": string;
}